# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Usuarios
  version: 1.0.0
servers:
- url: /
paths:
...
  /users:
    post:
      operationId: registerUser
      security: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Usuario registrado exitosamente
        "400":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ErrorResponse'
                type: array
          description: Datos de entrada inválidos
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Correo ya registrado
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error interno del servidor
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Servicio no disponible
      summary: Registrar un nuevo usuario
      x-content-type: application/json
      x-accepts:
        - application/json
...
  /users/{id}:
    delete:
      operationId: deleteUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutAfterDeleteUsers'
          description: usuario eliminado correctamente
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: usuario no autorizado
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Prohibido (sin permisos suficientes)
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Usuario no encontrado
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error interno en el servidor
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Servicio no disponible
      summary: Eliminar Usuario
      x-accepts:
        - application/json
    get:
      operationId: getUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Usuario encontrado
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: usuario no autorizado
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Prohibido (sin permisos suficientes)
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Usuario no encontrado
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error interno en el servidor
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Servicio no disponible
      summary: Busqueda del Usuario Especificado
      x-accepts:
        - application/json
    put:
      operationId: updateUser
      parameters:
        - explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPUT'
        description: Datos del usuario a actualizar
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: usuario actualizado correctamente
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: usuario no autorizado
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Prohibido (sin permisos suficientes)
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Usuario no encontrado
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error interno en el servidor
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Servicio no disponible
      summary: Actualizar usuario
      x-content-type: application/json
      x-accepts:
        - application/json
...
components:
  schemas:
    UserRegistration:
      example:
        password: password
        fullName: fullName
        dateBirth: 1990-01-01
        email: email
        rol: user
      properties:
        email:
          format: email
          type: string
        password:
          description: "La contraseña del usuario (debe contener al menos un dígito,\
            \ una mayúscula y una minúscula)"
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          type: string
        fullName:
          maxLength: 100
          type: string
        dateBirth:
          example: 1990-01-01
          format: date
          type: string
        rol:
          default: user
          enum:
          - user
          - admin
          type: string
      required:
      - email
      - fullName
      - password
      type: object
...
    SuccessResponse:
      description: marcar accion realizada exitosamente
      example:
        message: Contraseña actualizada exitosamente
      properties:
        message:
          example: Contraseña actualizada exitosamente
          type: string
      required:
      - message
      type: object
    UserResponse:
      example:
        fullName: fullName
        id: id
        dateBirth: 2000-01-23
        email: email
        rol: rol
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        dateBirth:
          format: date
          type: string
        rol:
          type: string
      required:
      - email
      - fullName
      - id
      - rol
      type: object
    ErrorResponse:
      example:
        error: error
        message: message
      properties:
        error:
          type: string
        message:
          type: string
      required:
      - error
      - message
      type: object
    OutAfterDeleteUsers:
      example:
        message: El usuario se ha eliminado con exito
      properties:
        message:
          example: El usuario se ha eliminado con exito
          type: string
      type: object
    UpdateUserPUT:
      example:
        datebirth: 2000-01-23
        fullname: fullname
      properties:
        fullname:
          maxLength: 80
          type: string
        datebirth:
          format: date
          type: string
      required:
      - datebirth
      - fullname
      type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de autenticación JWT. Debe ser incluido en el encabezado de autorización de la solicitud.
        Ejemplo: `Authorization: Bearer <token>`